Extract notification times logic into lib/notification-times.ts
Moves the pure checkWithinNotificationTimes function to a file with no
Electron dependencies so the test file can import it directly without mocking.
